backend/requirements.txt
Django==5.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.1
python-decouple==3.8
qrcode==7.4.2
Pillow>=11.0.0
gunicorn==21.2.0
